Rename to airframe-codec
