--- msgpack/src/test/scala/wvlet/airframe/msgpack/codec/MessageCodecTest.scala
@@ -1,31 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package wvlet.airframe.msgpack.codec
-
-import wvlet.airframe.AirframeSpec
-import wvlet.airframe.msgpack.codec.PrimitiveCodec.LongCodec
-
-/**
-  *
-  */
-class MessageCodecTest extends AirframeSpec {
-  "MessageCodec" should {
-
-    "have surface" in {
-      info(LongCodec.surface)
-    }
-
-  }
-
-}
