+++ tablet/src/main/scala/wvlet/airframe/tablet/text/PrettyPrint.scala
@@ -1,6 +1,6 @@
 package wvlet.airframe.tablet.text
 
-import wvlet.airframe.msgpack.codec.MessageCodec
+import wvlet.airframe.codec.MessageCodec
 import wvlet.airframe.tablet.obj.ObjectTabletReader
 import wvlet.log.LogSupport
 import wvlet.surface.Surface
